Separate blast data
