+++ docs/index.html
@@ -31,7 +31,7 @@
    sizes="16x16"
    href="assets/img/favicon/favicon-16x16.png"
   />
-  <link rel="manifest" href="assets/favicon/manifest.json" />
+  <!--<link rel="manifest" href="assets/favicon/manifest.json" />-->
   <link
    rel="mask-icon"
    href="assets/img/favicon/safari-pinned-tab.svg"
